+++ test/chez/aspect-build-app/src/app/invalid_decl.clj
@@ -1,0 +1,5 @@
+(ns app.invalid-decl)
+
+(def ^{:jolt.aspects/id :test/invalid-declare
+       :jolt.aspects/role :test/entry-around}
+  invalid-declare)