+++ test/chez/aspect-build-app/src/app/invalid_value.clj
@@ -1,0 +1,6 @@
+(ns app.invalid-value)
+
+(def ^{:jolt.aspects/id :test/invalid-value
+       :jolt.aspects/role :test/entry-around}
+  invalid-value
+  1)