+++ test/chez/aspect-build-app/src/app/target.clj
@@ -6,7 +6,9 @@
     (throw (ex-info "application failure" {:kind :application}))
     (str x "!")))
 
-(defn callback [x]
+(defn ^{:jolt.aspects/id :test/callback-entry
+        :jolt.aspects/role :test/entry-around}
+  callback [x]
   (println (str "callback " x))
   (cond
     (= x "recur") (recur "done")
@@ -17,6 +19,8 @@
   ;; No resolved app.target/callback invocation exists for a call selector.
   (f x))
 
-(defn numeric-callback [^long x]
+(defn ^{:jolt.aspects/id :test/numeric-callback-entry
+        :jolt.aspects/role :test/numeric-entry-around}
+  numeric-callback [^long x]
   (println (str "numeric-callback " x))
   (inc x))